feat: add dynamic import load helper
