+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './match';
+export * from './load';
 export * from './use-routes';
 export * from './use-router';
 export { default as Router } from './router';
